CSS and FS refactoring/Finished dropdown
