+++ src/app/app.component.html
@@ -1,7 +1,7 @@
 <div class="main-wrapper">
     <div class="title-bar">
         <fa-icon [icon]="faChartColumn"></fa-icon>
-        <p class="title">PokeStats</p>
+        <h1 class="title">PokeStats</h1>
     </div>
