+++ src/app/app.module.ts
@@ -12,6 +12,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
+import { SharedModule } from './sharedModule/shared.module';
 
 @NgModule({
   declarations: [
@@ -30,10 +31,15 @@ import { MatInputModule } from '@angular/material/input';
     FontAwesomeModule,
     MatFormFieldModule,
     MatInputModule,
-    ReactiveFormsModule
+    ReactiveFormsModule,
+    SharedModule,
   ],
-  providers: [],
-  bootstrap: [AppComponent]
+  providers: [
+
+  ],
+  bootstrap: [
+    AppComponent
+  ]
 })
 
 export class AppModule {
