+++ src/app/components/pokemon-card/pokemon-card.component.html
@@ -1,7 +1,7 @@
 <div class="main-wrapper">
     <div class="title">
-        <p>Name: {{currentPokemon.name | titlecase}}</p>
-        <p>Pokedex #{{currentPokemon.id}}</p>
+        <p class="section-header pokemonNumber">No.{{currentPokemon.id}}</p>
+        <p class="section-header pokemonName">{{currentPokemon.name | removeHyphen | titlecase }}</p>
     </div>
