+++ src/app/components/pokemon-card/pokemon-card.component.scss
@@ -2,11 +2,15 @@
 
 .title,
 .sprite-image {
-    background-color: $dark-yellow;
+    background: radial-gradient(circle, rgba(90,112,130,1) 30%, #4b5d6c 100%);
     border-radius: 5px;
     border-style: solid;
     border-width: 2px;
     border-color: $bright-yellow;
+}
+
+.pokemonNumber {
+    width: 35%;
 }
 
 .sprite-image {
@@ -29,9 +33,7 @@
 
 .title {
     padding: 20px;
-    p {
-        color: $white;
-        font-size: 24px;
-        margin: 0;
-    }
+    background: linear-gradient(135deg, $dark-blue-grey 30%, $blue-grey 30%);
+    display: flex;
+    flex-direction: row;
 }
