+++ src/app/components/search-bar/search-bar.component.scss
@@ -69,7 +69,7 @@
     display: flex;
     min-width: 8%;
     justify-content: center;
-    background-color: $dark-yellow;
+    background-color: $blue-grey;
     border-radius: 50%;
     align-items: center;
 }
