+++ src/app/components/search-bar/search-bar.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { PokemonService } from 'src/app/services/pokemon.service';
+import { PokemonService } from 'src/app/sharedModule/services/pokemon.service';
 import { faMagnifyingGlass, faDice } from '@fortawesome/free-solid-svg-icons';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { HttpClient } from '@angular/common/http';
