+++ src/app/components/stat-card/stat-card.component.scss
@@ -7,10 +7,26 @@
 }
 
 .card-section {
-    background-color: $dark-yellow;
+    background-color: $blue-grey;
     padding: 20px;
     border-radius: 5px;
     border-style: solid;
     border-color: $bright-yellow;
     border-width: 2px;
+}
+
+.ability-list-wrapper {
+    display: flex;
+}
+
+.abilityChip {
+    padding: 5px;
+}
+
+.hiddenAbilites {
+    background-color: $hidden-ability;
+}
+
+.regularAbilites {
+    background-color: $normal-ability;
 }
