+++ src/app/sharedModule/pipes/remove-hyphen.pipe.ts
@@ -1,0 +1,13 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+@Pipe({
+  name: 'removeHyphen'
+})
+
+export class RemoveHyphenPipe implements PipeTransform {
+
+  transform(value: string): string {
+    return value.replace(/-/g, ' ');
+  }
+
+}
